sefaz_xml_downloader/requirements.txt
requests
requests-pkcs12
lxml
python-dotenv
psycopg2